organized the project folders and packages and refactored the databuilder code and its respective classes
