+++ src/test/java/com/restfulbooker/RestfulBookerE2ETests.java
@@ -1,33 +1,34 @@
 package com.restfulbooker;
 
+import static data.restfulbooker.BookingDataBuilder.*;
+import static data.restfulbooker.TokenBuilder.getToken;
 import static io.restassured.RestAssured.given;
 import static org.hamcrest.Matchers.*;
 
+import data.restfulbooker.BookingData;
+import data.restfulbooker.PartialBookingData;
+import data.restfulbooker.Tokencreds;
 import org.hamcrest.Matchers;
 import org.testng.annotations.BeforeTest;
 import org.testng.annotations.Test;
-
-import data.*;
 
 /**
  * Created By Faisal Khatri on 18-02-2022
  */
 public class RestfulBookerE2ETests extends BaseSetup {
 
-    private BookingData        newBooking;
+    private BookingData newBooking;
     private BookingData        updatedBooking;
     private PartialBookingData partialUpdateBooking;
-    private Tokencreds         tokenCreds;
+    private Tokencreds tokenCreds;
     private int                bookingId;
 
     @BeforeTest
     public void testSetup () {
-        BookingDataBuilder builder = new BookingDataBuilder ();
-        TokenBuilder tokenbuild = new TokenBuilder ();
-        newBooking = builder.bookingDataBuilder ();
-        updatedBooking = builder.bookingDataBuilder ();
-        partialUpdateBooking = builder.partialBookingBuilder ();
-        tokenCreds = tokenbuild.tokenBuilder ();
+        newBooking = getbookingData();
+        updatedBooking = getbookingData ();
+        partialUpdateBooking = getpartialBookingData();
+        tokenCreds = getToken();
     }
 
     @Test
