+++ src/test/java/data/restfulbooker/BookingData.java
@@ -1,12 +1,12 @@
-package data;
+package data.restfulbooker;
 
 import lombok.Builder;
-import lombok.Data;
+import lombok.Getter;
 
 /**
  * Created By Faisal Khatri on 18-02-2022
  */
-@Data
+@Getter
 @Builder
 public class BookingData {
     private String firstname;
