+++ src/test/java/data/restfulbooker/BookingDates.java
@@ -1,14 +1,15 @@
-package data;
+package data.restfulbooker;
 
 import lombok.Builder;
 import lombok.Data;
+import lombok.Getter;
 
 import java.util.Date;
 
 /**
  * Created By Faisal Khatri on 18-02-2022
  */
-@Data
+@Getter
 @Builder
 public class BookingDates {
    private  String checkin;
