+++ src/test/java/data/restfulbooker/PartialBookingData.java
@@ -1,12 +1,13 @@
-package data;
+package data.restfulbooker;
 
 import lombok.Builder;
 import lombok.Data;
+import lombok.Getter;
 
 /**
  * Created By Faisal Khatri on 19-02-2022
  */
-@Data
+@Getter
 @Builder
 public class PartialBookingData {
     String firstname;
