+++ src/test/java/data/restfulbooker/TokenBuilder.java
@@ -1,11 +1,11 @@
-package data;
+package data.restfulbooker;
 
 /**
  * Created By Faisal Khatri on 19-02-2022
  */
 public class TokenBuilder {
 
-    public Tokencreds tokenBuilder () {
+    public static Tokencreds getToken () {
         return Tokencreds.builder ()
                 .username ("admin")
                 .password ("password123")
