+++ src/test/java/data/restfulbooker/Tokencreds.java
@@ -1,13 +1,14 @@
-package data;
+package data.restfulbooker;
 
 import lombok.Builder;
 import lombok.Data;
+import lombok.Getter;
 
 /**
  * Created By Faisal Khatri on 19-02-2022
  */
 @Builder
-@Data
+@Getter
 public class Tokencreds {
 
     private String username;
