+++ test-suite/restfulbookersuite.xml
@@ -3,7 +3,7 @@
 <suite name="Restful Booker Test Suite">
     <listeners>
         <listener
-                class-name="io.github.mfaisalkhatri.TestListener"/>
+                class-name="io.github.mfaisalkhatri.utility.TestListener"/>
     </listeners>
     <test name="Restful Booker tests">
         <classes>
