+++ test-suite/testng.xml
@@ -3,7 +3,7 @@
 <suite name="Rest-Assured API Test Suite">
     <listeners>
         <listener
-                class-name="io.github.mfaisalkhatri.TestListener"/>
+                class-name="io.github.mfaisalkhatri.utility.TestListener"/>
     </listeners>
     <suite-files>
             <suite-file path="reqrestestsuite.xml"></suite-file>
